test(integration): fix test recordings for embeds on universal runtime
adobe/helix-pipeline#1056 switches the embed service URL to universal runtime. This means that some of the recordings for the integration tests need to be re-downloaded, as the URL has changed. As helix-embed has evolved in the meantime, this also means that the test output has changed a bit. There is a good chance that we will end up in a smoke test deadlock, as this PR won't pass the tests until adobe/helix-pipeline#1056 has been merged and vice versa.