feat(security): provide more troubleshooting instructions on `aem up`
The `unable to get local issuer certificate` error has been stumping our customers for a while now
and it typically takes some time to route the issue to me, who can point to the nodejs docs, customer's
IT department, and firewall. By intercepting the error code I hope that we can point people into the
right direction a bit faster.